requirements.txt
flask
textblob
python-dotenv
google-generativeai
